Created 1st Battle Message
For the time being, I'm rendering the battle messages in the top left area of the screen. I only print "A Hostile Robot Leader has appeared!" at the start of the battle. I will now insert a black rectangle within a white rectangle, which will serve as the dialogue box where the battle messages will be rendered.