DEBUG: Rendering messages based on Keys
I'm now rendering messages on the console if the user presses "1", "2", or "3" on the keyboard while the Battle Menu is being displayed. If they press "1", I print a message in the console saying that they are attacking. If they press "2", I render that they are on guard. Finally, if they press "3", I render a message on the console with a print() statement saying that the player drank a potion.