Rendering Player Attack (BUGGY)
The good thing is that I’m rendering the names of the characters and the damage dealt by the player by using formatted strings and by calling the Character class by calling the “player” and “enemy” variables. The bad thing is that, if the message is too long, the message will overflow, and the user won’t be able to read the entire message.