Rendering Enemy's HP in UI
I'm rendering the enemy's HP in the game's UI. I had to put a black box with a white rectangle behind the enemy's HP in the UI to hide the UI text when the enemy was at full health, so that the text looks good in the UI. I also made other fixes, such as that now I'm rendering the player and the enemy's total HP without hard coding it. To do so, I had to create a new property in the character's class to store the characters total HP, which will never decrease.