Rendering Victory Message (BUGGY)
If the enemy’s HP is less or equal to 0, you get a victory message. The bad thing is that, just like when you get a game over, you can see that the UI with the enemy’s HP is displaying negative numbers, when it should render “0” if the enemy is defeated. Also, right after the “the player attacks” message, I never display how much damage points the player dealt to the enemy. Instead, I go to the victory message immediately after the “the player attacks!” message. Also, the game goes on and renders the battle menu once again if I press “z” after the victory message