BUGGY: Updating Numer of Potions
I’m now correctly updating the number of potions on the battle menu, so that, if I drink 1, it will be deducted from the total number of potions. I'm also healing the player if they use a potion by calling the Character class. The bad thing is that the bug that I predicted that would occur occurred. If I drink a potion, and it heals me more than my max amount of HP, my current HP will be higher than my total amount of HP, so my HP is higher than my max HP. And if I press “Z” after that, the player attacks. I shouldn’t attack right after healing!