BUGFIX: Fixed Potion Mechanic
The bug that caused the player to attack right after drinking a potion was that
the "has player attacked" boolean was never being disabled after attacking. Also, I fixed the bug in which the game
froze if I drank a potion on the very first turn. The bug was that the player's turn never ended after drinking a potion. So, I fixed 2 bugs in total.