Created First Function: Damage
I created a damage dealing function, which will calculate a random number to assign the total number of attack points from the player or the enemy before returning the total number of damage points that the character should deal. I'm only using it for the player for the time being. I need to assign it to the enemy as well.