HP Remaining() function Created
I created a function that will calculate how many health points the character has left (be it either the enemy
or the player). It will return an integer, which will be the new number that should be added to the enemy or to
the player's health points. Sure, the hp remaining() function is just a subtraction, but it also makes the HP
automatically zero if the HP is negative. Neither the player nor the enemy should show negative HP. The function works perfectly, and the game still correctly gives a game over or a victory screen if the enemy or the player is defeated, respectively.