Passing First Unit Test
The Damage Calculation() function raises a Value Error if you insert a non-integer or a negative number into it. I also created my first unit test,s which tests the damage calculation function. If it detects that you try to pass a negative number, it raises a Value Error, so the test passes.