Passing Second Unit Test
The HP Remaining Calculation() function raises a Value Error if none of the 2 parameters that you need to insert to it is an integer. Then, I created a unit test that raises a Value Error if the first parameter of the HP Remaining Calculation() function is a float. The test passes.